refactor: :recycle: move compare branches functionality to reusable util function
