+++ src/main/api/gitHub/compareBranches.ts
@@ -1,13 +1,6 @@
-import { RestEndpointMethodTypes } from '@octokit/rest';
-import { ReadCommitResult } from 'isomorphic-git';
-import path from 'path';
-import { mainStore as store } from 'src/main';
-import { MAIN_BRANCH } from '../../../shared/constants';
-import { activePublication } from '../../../shared/redux/slices/loadPublicationsSlice';
-import { LocalPublication } from '../../../shared/types';
-import { BranchComparison, IpcEventHandler } from '../../../shared/types/api';
-import createGitHubHandler from '../../lib/gitHubHandler';
-import createGitRepoHandler from '../../lib/gitRepoHandler';
+import { BranchComparison } from '../../../shared/types';
+import { IpcEventHandler } from '../../../shared/types/api';
+import compareBranchesUtil from '../../utils/git/compareBranches';
 
 const compareBranches: IpcEventHandler<BranchComparison> = async (
   _,
@@ -15,55 +8,8 @@ const compareBranches: IpcEventHandler<BranchComparison> = async (
   branchB?: string,
   opts?: { useRemote?: boolean }
 ) => {
-  const publication = activePublication(store.getState()) as LocalPublication;
-  const token = store.getState().currentUser.auth.accessToken?.value;
-  if (!publication) {
-    throw new Error('No active publication!');
-  }
-  if (!token) {
-    throw new Error('No user is logged in!');
-  }
-
-  const repoHandler = createGitRepoHandler(publication);
-  const githubHandler = createGitHubHandler(token);
-  const [commitsA, commitsB] = await Promise.all([
-    (await repoHandler.log(branchA)).map(normalizeCommitData),
-    (opts?.useRemote
-      ? await githubHandler.getCommits(
-          {
-            owner: publication.owner,
-            name: path.basename(publication.dirPath),
-          },
-          branchB
-        )
-      : await repoHandler.log(branchB || MAIN_BRANCH)
-    ).map(normalizeCommitData),
-  ]);
-
-  const result = {
-    isAhead: commitsB.some(
-      (commitB) => !commitsA.some((commitA) => commitA.sha === commitB.sha)
-    ),
-    isBehind: commitsA.some(
-      (commitA) => !commitsB.some((commitB) => commitA.sha === commitB.sha)
-    ),
-  };
+  const result = await compareBranchesUtil(branchA, branchB, opts);
   return result;
 };
 
 export default compareBranches;
-
-function normalizeCommitData(
-  data:
-    | ReadCommitResult
-    | RestEndpointMethodTypes['repos']['listCommits']['response']['data'][0]
-) {
-  if ('oid' in data) {
-    return {
-      sha: data.oid,
-    };
-  }
-  return {
-    sha: data.sha,
-  };
-}
